backend/requirements.txt

Flask
flask-cors
pymongo
python-dotenv
openai
    